Homework2 (Part A)
Write a program that reads dates from input, one date per line. Each
date's format must be as follows: March 1, 1990. Any date not following
that format is incorrect and should be ignored. Use the find() method to
parse the string and extract the date. The input ends with -1 on a line
alone. Use the datetime module to get the current date. Ignore any dates
that are later than the current date. Output each correct date as:
3/1/1990.